+++ server/ver1_0/BookMarkGroupServer/Model/groupFolderModel.py
@@ -1,32 +1,43 @@
-import db
+from module import baseDB
+import numberingModel
 
-class AuthorityModel(db.Base_db):
+class GroupFolderModel(baseDB.BaseDB):
 	def __init__(self):
 		self.table_info = {
-			"table": "authority",
+			"table": "group_folder",
 			"param": {
-				"user_id": { "primary_key":"YES", "insert": "YES", "update": "NO"},
 				"group_folder_id": { "primary_key":"YES", "insert": "YES", "update": "NO"},
-				"authority_state" : { "primary_key":"NO", "insert": "YES", "update": "NO"},
+				"group_folder_name": { "primary_key":"NO", "insert": "YES", "update": "NO"},
+				"group_folder_version": { "primary_key":"NO", "insert": "YES", "update": "YES"},
+				"json_directory_path": { "primary_key":"NO", "insert": "YES", "update": "NO"},
+				"group_folder_memo": { "primary_key":"NO", "insert": "YES", "update": "YES" },
 				"create_time": { "primary_key":"NO" , "insert": "YES", "update": "NO"},
 				"update_time": { "primary_key":"NO" , "insert": "YES", "update": "YES"},
 			}
 		}
 
-	def selectAuthority(self, request):
+	def selectGroupFolder(self, request):
 		# Create select_value from requested data
 		select_value = self.createRequestSqlValue(request)
 
-		# Create sql statement(This function is override)
+		# Create sql statement
 		select_sql = self.createSelectSql(self.table_info)
 
 		# Perform processing
 		select_data, message = self.selectExecute(select_sql, select_value)
 		return select_data, message
 
-	def insertAuthority(self, request):
+	def insertGroupFolder(self, request):
+		
+		numbering_class = numberingModel.NumberingModel()
+
 		# Create insert_value from requested data
 		insert_value = self.createRequestSqlValue(request)
+		# Acquire group_folder_id from numbering DB
+		insert_value["group_folder_id"] = numbering_class.getNextValue("group_folder_id")
+		insert_value["group_folder_version"] = 1
+		insert_value["json_directory_path"] = insert_value["group_folder_id"] + "_" + insert_value["group_folder_name"]
+		insert_value["group_folder_memo"] = ""
 		insert_value["create_time"] = self.create_time()
 		insert_value["update_time"] = self.create_time()
 
@@ -37,7 +48,8 @@ class AuthorityModel(db.Base_db):
 		message = self.insertExecute(insert_sql, insert_value)
 		return message
 
-	def upadteAuthority(self, request):
+	def updateGroupFolder(self, request):
+
 		# Create update_value from requested data
 		update_value = self.createRequestSqlValue(request)
 		update_value["update_time"] = self.create_time()
@@ -49,7 +61,7 @@ class AuthorityModel(db.Base_db):
 		message = self.updateExecute(update_sql, update_value)
 		return message
 
-	def delateAuthority(self, request):
+	def delateGroupFolder(self, request):
 		# Create delate_value from requested data
 		delate_value = self.createRequestSqlValue(request)
 
@@ -59,6 +71,3 @@ class AuthorityModel(db.Base_db):
 		# Perform processing
 		message = self.delateExecute(delate_sql, delate_value)
 		return message
-
-	def createSelectSql(self, request):
-		return ""
