+++ server/ver1_0/BookMarkGroupServer/Model/userModel.py
@@ -1,22 +1,21 @@
-import db
+from module import baseDB
 import numberingModel
 
-class GroupFolderModel(db.Base_db):
+class UserModel(baseDB.BaseDB):
 	def __init__(self):
 		self.table_info = {
-			"table": "group_folder",
+			"table": "user",
 			"param": {
-				"group_folder_id": { "primary_key":"YES", "insert": "YES", "update": "NO"},
-				"group_folder_name": { "primary_key":"NO", "insert": "YES", "update": "NO"},
-				"group_folder_version": { "primary_key":"NO", "insert": "YES", "update": "YES"},
-				"json_directory_path": { "primary_key":"NO", "insert": "YES", "update": "NO"},
-				"group_folder_memo": { "primary_key":"NO", "insert": "YES", "update": "YES" },
-				"create_time": { "primary_key":"NO" , "insert": "YES", "update": "NO"},
-				"update_time": { "primary_key":"NO" , "insert": "YES", "update": "YES"},
+				"user_id": {"primary_key":"YES", "insert": "YES", "update": "NO"},
+				"user_name": {"primary_key":"NO", "insert": "YES", "update": "YES"},
+				"user_mail_address": {"primary_key":"NO", "insert": "YES", "update": "YES"},
+				"user_password":  {"primary_key":"NO", "insert": "YES", "update": "YES"},
+				"create_time": { "primary_key":"NO", "insert": "YES", "update": "NO"},
+				"update_time": { "primary_key":"NO", "insert": "YES", "update": "YES"},
 			}
 		}
 
-	def selectGroupFolder(self, request):
+	def selectUser(self, request):
 		# Create select_value from requested data
 		select_value = self.createRequestSqlValue(request)
 
@@ -27,17 +26,15 @@ class GroupFolderModel(db.Base_db):
 		select_data, message = self.selectExecute(select_sql, select_value)
 		return select_data, message
 
-	def insertGroupFolder(self, request):
-		
+	def insertUser(self, request):
 		numbering_class = numberingModel.NumberingModel()
 
 		# Create insert_value from requested data
 		insert_value = self.createRequestSqlValue(request)
 		# Acquire group_folder_id from numbering DB
-		insert_value["group_folder_id"] = numbering_class.getNextValue("group_folder_id")
-		insert_value["group_folder_version"] = 1
-		insert_value["json_directory_path"] = insert_value["group_folder_id"] + "_" + insert_value["group_folder_name"]
-		insert_value["group_folder_memo"] = ""
+		insert_value["group_folder_id"] = numbering_class.getNextValue("user_id")
+		# Encrypt password 
+		#insert_value["user_mail_address"] = 
 		insert_value["create_time"] = self.create_time()
 		insert_value["update_time"] = self.create_time()
 
@@ -48,10 +45,14 @@ class GroupFolderModel(db.Base_db):
 		message = self.insertExecute(insert_sql, insert_value)
 		return message
 
-	def updateGroupFolder(self, request):
-
+	def updateUser(self, request):
 		# Create update_value from requested data
 		update_value = self.createRequestSqlValue(request)
+
+		if request.form["user_password"] != "password":
+			# Encrypt password 
+			#update_value["user_mail_address"] = 
+			pass
 		update_value["update_time"] = self.create_time()
 
 		# Create sql statement
@@ -60,8 +61,8 @@ class GroupFolderModel(db.Base_db):
 		# Perform processing
 		message = self.updateExecute(update_sql, update_value)
 		return message
-
-	def delateGroupFolder(self, request):
+	
+	def delateNumbergin(self, request):
 		# Create delate_value from requested data
 		delate_value = self.createRequestSqlValue(request)
