+++ server/ver1_0/BookMarkGroupServer/main.py
@@ -1,5 +1,6 @@
 from flask import Flask, render_template, redirect, url_for
 from flask_login import LoginManager, login_url, login_required
+from module.BookMarkGroupController import BookMarkGroupController
 
 app = Flask(__name__)
 app.secret_key = 'BookMarkGroup Secret Key'
@@ -7,72 +8,80 @@ app.secret_key = 'BookMarkGroup Secret Key'
 login_maneger = LoginManager()
 login_maneger.init_app(app)
 
-#Root
+bookMarkGroup_class = BookMarkGroupController.BookMarkGroupController()
+
+# Root
 @app.route('/', methods=['GET'])
 def root():
 	return redirect(url_for('g_loginUser'))
 
-#Create User
-@app.route('/createUser', methods=['GET'])
-def g_createUser():
-	pass
-@app.route('/createUser', methods=['POST'])
-def p_createUser():
-	pass
+# Create User
+@app.route('/createUser', methods=['GET', 'POST'])
+def createUser():
+	bookMarkGroup_class.createUser()
+	return
 
-#Login User
-@app.route('/loginUser', methods=['GET'])
-def g_loginUser():
-	pass
-@app.route('/loginUser', methods=['POST'])
-def p_loginUser():
-	pass
+# Login User
+@app.route('/loginUser', methods=['GET', 'POST'])
+def loginUser():
+	bookMarkGroup_class.loginUser()
+	return
 
-#Sync Folders
+# Sync Folders
 @app.route('/syncFolders', methods=['GET'])
-def g_syncFolders():
-	pass
+def syncFolders():
+	bookMarkGroup_class.syncFolders()
+	return
 
-#Create GroupFolder
+# Create GroupFolder
 @app.route('/createGroupFolder', methods=['POST'])
-def p_createGroupFolder():
-	pass
+def createGroupFolder():
+	bookMarkGroup_class.createGroupFolder()
+	return
 
-#Commit GroupFolder
+# Commit GroupFolder
 @app.route('/commitGroupFolder', methods=['POST'])
-def p_commitGroupFolder():
-	pass
+def commitGroupFolder():
+	bookMarkGroup_class.commitGroupFolder()
+	return
 
-#Update GroupFolder
+# Update GroupFolder
 @app.route('/updateGroupFolder', methods=['POST'])
-def p_updateGroupFloder():
-	pass
+def updateGroupFloder():
+	bookMarkGroup_class.updateGroupFolder()
+	return
 
-#Delete GroupFolder
+# Delete GroupFolder
 @app.route('/deleteGroupFolder', methods=['POST'])
-def p_deleteGroupFolder():
-	pass
+def deleteGroupFolder():
+	bookMarkGroup_class.deleteGroupFolder()
+	return
 
-#Setting GroupFolder
+# Setting GroupFolder
 @app.route('/settingShowUser', methods=['POST'])
-def p_settingShowUser():
-	pass
+def settingShowUser():
+	bookMarkGroup_class.settingShowUser()
+	return
 
 #Setting GroupFolder (Add user)
 @app.route('/settingAddUser', methods=['POST'])
-def p_settingAddUser():
-	pass
+def settingAddUser():
+	bookMarkGroup_class.settingAddUser()
+	return
 	
 #Setting GroupFolder (Del usr)
 @app.route('/settginDelUser', methods=['POST'])
-def p_settingDelUser():
-	pass
+def settingDelUser():
+	bookMarkGroup_class.settingDelUser()
+	return
 
 #Setting GroupFolder (Setting Authority)
 @app.route('/settingSetAuthority', methods=['POST'])
-def p_settingSetAuthority():
-	pass
+def settingSetAuthority():
+	bookMarkGroup_class.settingSetAuthority()
+	return
 
 @app.route('/settingShowAuthority', methods=['POST'])
-def p_settingShowAuthority():
-	pass
+def settingShowAuthority():
+	bookMarkGroup_class.settingShowAuthority()
+	return
